Correcting code with SonarLint(95/100)

+++ src/main/java/inventory/controller/MainScreenController.java
@@ -20,18 +20,15 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.Optional;
 import java.util.ResourceBundle;
+import java.util.logging.Logger;
 
 
 public class MainScreenController implements Initializable,Controller {
     
      // Declare fields
-    private Stage stage;
-    private Parent scene;
-    private static Part modifyPart;
-    private static Product modifyProduct;
     private static int modifyPartIndex;
     private static int modifyProductIndex;
-    
+    private Logger logger = Logger.getLogger(MainScreenController.class.getName());
     // Declare methods
     public static int getModifyPartIndex() {
         return modifyPartIndex;
@@ -80,8 +77,6 @@ public class MainScreenController implements Initializable,Controller {
     @FXML
     private TextField productsSearchTxt;
 
-    public MainScreenController(){}
-
     public void setService(InventoryService service){
         this.service=service;
         partsTableView.setItems(service.getAllParts());
@@ -114,10 +109,9 @@ public class MainScreenController implements Initializable,Controller {
      * @throws IOException
      */
     private void displayScene(ActionEvent event, String source) throws IOException {
-        stage = (Stage)((Button)event.getSource()).getScene().getWindow();
+        Stage stage = (Stage)((Button)event.getSource()).getScene().getWindow();
         FXMLLoader loader= new FXMLLoader(getClass().getResource(source));
-        //scene = FXMLLoader.load(getClass().getResource(source));
-        scene = loader.load();
+        Parent scene = loader.load();
         Controller ctrl=loader.getController();
         ctrl.setService(service);
         stage.setScene(new Scene(scene));
@@ -139,11 +133,11 @@ public class MainScreenController implements Initializable,Controller {
         alert.setContentText("Are you sure you want to delete part " + part.getName() + " from parts?");
         Optional<ButtonType> result = alert.showAndWait();
 
-        if (result.get() == ButtonType.OK) {
-            System.out.println("Part deleted.");
+        if(result.isPresent() && result.get() == ButtonType.OK) {
+            logger.info("Part deleted.");
             service.deletePart(part);
         } else {
-            System.out.println("Canceled part deletion.");
+            logger.info("Canceled part deletion.");
         }
     }
 
@@ -161,12 +155,12 @@ public class MainScreenController implements Initializable,Controller {
         alert.setHeaderText("Confirm Product Deletion?");
         alert.setContentText("Are you sure you want to delete product " + product.getName() + " from products?");
         Optional<ButtonType> result = alert.showAndWait();
-        
-        if (result.get() == ButtonType.OK) {
+
+        if(result.isPresent() && result.get() == ButtonType.OK) {
             service.deleteProduct(product);
-            System.out.println("Product " + product.getName() + " was removed.");
+            logger.info(String.format("Product {} was removed.", product.getName()));
         } else {
-            System.out.println("Product " + product.getName() + " was not removed.");
+            logger.info(String.format("Product {} was not removed.", product.getName()));
         }
     }
 
@@ -198,7 +192,7 @@ public class MainScreenController implements Initializable,Controller {
      */
     @FXML
     void handleModifyPart(ActionEvent event) throws IOException {
-        modifyPart = partsTableView.getSelectionModel().getSelectedItem();
+        Part modifyPart = partsTableView.getSelectionModel().getSelectedItem();
         modifyPartIndex = service.getAllParts().indexOf(modifyPart);
         
         displayScene(event, "/fxml/ModifyPart.fxml");
@@ -211,7 +205,7 @@ public class MainScreenController implements Initializable,Controller {
      */
     @FXML
     void handleModifyProduct(ActionEvent event) throws IOException {
-        modifyProduct = productsTableView.getSelectionModel().getSelectedItem();
+        Product modifyProduct = productsTableView.getSelectionModel().getSelectedItem();
         modifyProductIndex = service.getAllProducts().indexOf(modifyProduct);
         
         displayScene(event, "/fxml/ModifyProduct.fxml");
@@ -229,11 +223,11 @@ public class MainScreenController implements Initializable,Controller {
         alert.setHeaderText("Confirm Exit");
         alert.setContentText("Are you sure you want to exit?");
         Optional<ButtonType> result = alert.showAndWait();
-        if(result.get() == ButtonType.OK) {
-            System.out.println("Ok selected. Program exited");
+        if(result.isPresent() && result.get() == ButtonType.OK) {
+            logger.info("Ok selected. Program exited");
             System.exit(0);
         } else {
-            System.out.println("Cancel clicked.");
+            logger.info("Cancel clicked.");
         }
     }
